Move image cleanup/persistence to context
